Automatic room placement, multiple dungeon floors
Added generation of floor maps to randomly place generated rooms also created system to place a ladder to the next floor 
that regenerates the map and places the player back in middle of room. map and drawing hit-boxes can be turned on and off by named variables near the top.